changed: Appbar to own component
