+++ src/pages/App.tsx
@@ -1,7 +1,7 @@
 import "../styles/global.scss"
 
 import React, { ReactElement, useCallback } from "react"
-import { Route, Switch, Link, useLocation } from "react-router-dom"
+import { Route, Switch } from "react-router-dom"
 
 import { AppDispatch } from "../state"
 import { BLOCK_TIME } from "../constants"
@@ -15,42 +15,9 @@ import fetchGasPrices from "../utils/updateGasPrices"
 import fetchTokenPricesUSD from "../utils/updateTokenPrices"
 import { useDispatch } from "react-redux"
 import usePoller from "../hooks/usePoller"
-import { makeStyles, withStyles } from "@material-ui/core/styles"
-import AppBar from "@material-ui/core/AppBar"
-import Tabs from "@material-ui/core/Tabs"
-import Tab from "@material-ui/core/Tab"
-import Toolbar from "@material-ui/core/Toolbar"
-import IconButton from "@material-ui/core/IconButton"
-import { useTranslation } from "react-i18next"
-import logo from "../assets/icons/logo.svg"
-import Web3Status from "../components/material/Web3Status"
-import ThemeChanger from "../components/material/ThemeChanger"
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    alignItems: "stretch",
-  },
-  menuButton: {
-    marginRight: theme.spacing(2),
-  },
-}))
-
-const StyledTabs = withStyles({
-  root: {
-    flexGrow: 1,
-    "& div.MuiTabs-scroller": {
-      display: "inline-flex",
-      justifyContent: "center",
-    },
-  },
-})(Tabs)
+import Appbar from "../components/material/Appbar"
 
 export default function App(): ReactElement {
-  const classes = useStyles()
-  const { t } = useTranslation()
-  const location = useLocation()
-  const currentTab = location.pathname
-
   const dispatch = useDispatch<AppDispatch>()
 
   const fetchAndUpdateGasPrice = useCallback(() => {
@@ -65,36 +32,7 @@ export default function App(): ReactElement {
   return (
     <Web3ReactManager>
       <ToastsProvider>
-        <AppBar position="static">
-          <Toolbar className={classes.root}>
-            <IconButton
-              edge="start"
-              className={classes.menuButton}
-              color="inherit"
-              aria-label="menu"
-            >
-              <img className="logo" alt="logo" src={logo} />
-            </IconButton>
-            <StyledTabs value={currentTab} centered scrollButtons="off">
-              <Tab label={t("swap")} component={Link} to="/" value="/" />
-              <Tab
-                label={t("deposit")}
-                component={Link}
-                to="deposit"
-                value="/deposit"
-              />
-              <Tab
-                label={t("withdraw")}
-                component={Link}
-                to="withdraw"
-                value="/withdraw"
-              />
-              <Tab label={t("risk")} component={Link} to="risk" value="/risk" />
-            </StyledTabs>
-            <Web3Status />
-            <ThemeChanger />
-          </Toolbar>
-        </AppBar>
+        <Appbar />
         <Switch>
           <Route exact path="/" component={SwapBTC} />
           <Route exact path="/deposit" component={DepositBTC} />
